refactor: replace complex table with clean simple design
- Removed old FileVersionsTable component completely
- Created new DefinitionsTable with clean, minimal styling
- Simple border-based design matching the provided example
- No shadows, backgrounds, or complex containers
- Clean typography with proper spacing
- Improved date formatting (07 Aug 2025 format)
- Maintains existing colors (blue for links/cloud names)
- Eliminates all white bar spacing issues
